django_1/Grantha/requirements.txt
asgiref==3.2.3
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-crispy-forms==1.8.1
django-csvimport==2.13
django-import-export==2.0.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
PyYAML==5.3
sqlparse==0.3.0
tablib==1.1.0
views==0.3
xlrd==1.2.0
xlwt==1.3.0
